Move CommandAPI into CommandBean
